move luts out of image_config and to context
